+++ src/components/UserDetail/UserDetail.js
@@ -38,12 +38,12 @@ class UserDetail extends React.Component {
     openAccounts: {
       type: 'okapi',
       records: 'accounts',
-      path: 'accounts?query=(userId=!{user.id} and status.name<>Closed)&limit=100',
+      path: 'accounts?query=(userId==!{user.id} and status.name<>Closed)&limit=100',
     },
     openRequests: {
       type: 'okapi',
       throwErrors: false,
-      path: 'circulation/requests?query=(requesterId=!{user.id} and status=Open)&limit=100',
+      path: 'circulation/requests?query=(requesterId==!{user.id} and status=Open)&limit=100',
     },
   });