+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Updated React-intl dependency to 4.7.2
 * Use `==` for more efficient queries. Refs PERF-62.
 * pass proxy borrower's barcode to checkout endpoint. Fixes UICHKOUT-639
+* Use right truncation to fetch user's open request (UICHKOUT-641).
 
 ## [4.0.1](https://github.com/folio-org/ui-checkout/tree/v4.0.1) (2020-06-19)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v4.0.0...v4.0.1)