+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change history for ui-checkout
 
+## [4.0.2](https://github.com/folio-org/ui-checkout/tree/v4.0.2) (2020-07-10)
+[Full Changelog](https://github.com/folio-org/ui-checkout/compare/v4.0.1...v4.0.2)
+
+* pass proxy borrower's barcode to checkout endpoint. Fixes UICHKOUT-639
+
 ## [4.0.1](https://github.com/folio-org/ui-checkout/tree/v4.0.1) (2020-06-19)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v4.0.0...v4.0.1)